+++ HISTORY.md
@@ -1,7 +1,12 @@
+## 0.0.7
+
+- 
+
 ## 0.0.6
 
 - New modality switch added for rad/path
 - Added PyVIPS support
+- SubjectID-based split added
 
 ## 0.0.5